utils: cleanup
Move some RRM utility functions around to simplify code structure / #includes.

[fix commit message]
