+++ src/test/test_storage.c
@@ -334,44 +334,6 @@ static int load_time(time_t* v, char* s)
     return ret;
 }
 
-static int get_rrm_mode_val(char mode);
-static int get_rrm_mode_val(char mode) {
-    switch (tolower(mode)) {
-        case 'a':
-            return WLAN_RRM_CAPS_BEACON_REPORT_ACTIVE;
-            break;
-        case 'p':
-            return WLAN_RRM_CAPS_BEACON_REPORT_PASSIVE;
-            break;
-        case 'b':
-        case 't':
-             return WLAN_RRM_CAPS_BEACON_REPORT_TABLE;
-             break;
-    }
-    return 0;
-}
-
-static int parse_rrm_mode(int *rrm_mode_order, const char *mode_string);
-static int parse_rrm_mode(int *rrm_mode_order, const char *mode_string) {
-    int len, mode_val;
-    int mask = 0, order = 0, pos = 0;
-
-    if (!mode_string)
-        mode_string = DEFAULT_RRM_MODE_ORDER;
-    len = strlen(mode_string);
-
-    while (order < __RRM_BEACON_RQST_MODE_MAX) {
-        if (pos >= len) {
-            rrm_mode_order[order++] = 0;
-        } else {
-            mode_val = get_rrm_mode_val(mode_string[pos++]);
-            if (mode_val && !(mask & mode_val))
-                mask |= (rrm_mode_order[order++] = mode_val);
-        }
-    }
-    return mask;
-}
-
 static int consume_actions(int argc, char* argv[], int harness_verbosity);
 
 static int consume_actions(int argc, char* argv[], int harness_verbosity)
